Removed old TimeTravel drawer

+++ OSX/MacGUI/C64Proxy.mm
@@ -857,7 +857,8 @@ struct CRTContainerWrapper { CRTContainer *crtcontainer; };
 // Snapshot storage
 - (void) setAutoSaveSnapshots:(bool)b { wrapper->c64->autoSaveSnapshots = b; }
 - (NSInteger) numAutoSnapshots { return wrapper->c64->numAutoSnapshots(); }
-- (NSInteger) autoMostRecent:(NSInteger)nr { return wrapper->c64->autoMostRecent(nr); }
+- (NSInteger) autoMostRecent:(NSInteger)nr {
+    return wrapper->c64->autoMostRecent((unsigned)nr); }
 - (NSData *)autoSnapshotData:(NSInteger)nr {
     Snapshot *snapshot = wrapper->c64->autoSnapshot((unsigned)nr);
     return [NSData dataWithBytes: (void *)snapshot->header()
@@ -875,7 +876,8 @@ struct CRTContainerWrapper { CRTContainer *crtcontainer; };
     wrapper->c64->restoreAutoSnapshot((unsigned)nr); }
 
 - (NSInteger) numUserSnapshots { return wrapper->c64->numUserSnapshots(); }
-- (NSInteger) userMostRecent:(NSInteger)nr { return wrapper->c64->userMostRecent(nr); }
+- (NSInteger) userMostRecent:(NSInteger)nr {
+    return wrapper->c64->userMostRecent((unsigned)nr); }
 - (NSData *)userSnapshotData:(NSInteger)nr {
     Snapshot *snapshot = wrapper->c64->userSnapshot((unsigned)nr);
     return [NSData dataWithBytes: (void *)snapshot->header()
